rimae_backend/requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Database
psycopg2-binary>=2.9.9

# Authentication
djangorestframework-simplejwt>=5.3.0

# File uploads
Pillow>=10.0.0
django-storages>=1.14.0

# Environment
python-dotenv>=1.0.0

# API Documentation
drf-yasg>=1.21.0

# Utils
django-filter>=23.5
